hopefully actually fix list issues

+++ src/main/java/com/doctor/vampiricageing/capabilities/HunterAgeingManager.java
@@ -218,7 +218,7 @@ public class HunterAgeingManager {
         if(!attributes.get(Attributes.ATTACK_DAMAGE).isEmpty() && (Helper.isVampire(event.getEntity()) || CapabilityHelper.isWerewolfCheckMod(event.getEntity()))) {
             Player hunterSource = (Player) sourceEntity;
             int age = VampiricAgeingCapabilityManager.getAge(hunterSource).map(ageCap -> ageCap.getAge()).orElse(0);
-            event.setAmount(event.getAmount() + HunterAgeingConfig.ageEnemyFactionDamageIncrease.get().get(age));
+            event.setAmount(event.getAmount() + HunterAgeingConfig.ageEnemyFactionDamageIncrease.get().get(age).floatValue());
 
         }
     }
@@ -236,7 +236,7 @@ public class HunterAgeingManager {
             int cumulativeAge = CapabilityHelper.getCumulativeTaintedAge(player);
 
             if(event.getSource().is(DamageTypes.ON_FIRE) || event.getSource().is(DamageTypes.IN_FIRE)) {
-                event.setAmount(event.getAmount() * HunterAgeingConfig.taintedFireDamageMultiplier.get().get(cumulativeAge));
+                event.setAmount(event.getAmount() * HunterAgeingConfig.taintedFireDamageMultiplier.get().get(cumulativeAge).floatValue());
             }
         }
 
@@ -249,7 +249,7 @@ public class HunterAgeingManager {
         }
 
         int age = VampiricAgeingCapabilityManager.getAge(player).map(ageCap -> ageCap.getAge()).orElse(0);
-        event.setAmount(Math.round((float)event.getAmount() / HunterAgeingConfig.xpGainReduction.get().get(age)));
+        event.setAmount(Math.round((float)event.getAmount() / HunterAgeingConfig.xpGainReduction.get().get(age).floatValue()));
     }
     @SubscribeEvent
     public static void breakSpeed(PlayerEvent.BreakSpeed event) {
@@ -263,7 +263,7 @@ public class HunterAgeingManager {
             event.setCanceled(true);
         }
         int age = VampiricAgeingCapabilityManager.getAge(event.getEntity()).map(ageCap -> ageCap.getAge()).orElse(0);
-        event.setNewSpeed(event.getOriginalSpeed() * HunterAgeingConfig.hunterMiningSpeedBonus.get().get(age));
+        event.setNewSpeed(event.getOriginalSpeed() * HunterAgeingConfig.hunterMiningSpeedBonus.get().get(age).floatValue());
 
     }
     @SubscribeEvent
